source/requirements.txt
Werkzeug<1
Flask==1.0.2
PyMySQL==0.9.3
Flask-SQLAlchemy==2.4.4
Flask-Seeder==1.1.1
Flask-WTF==0.14.3
Flask-Login==0.5.0
email-validator==1.1.1
smallthon==1.2
python-dotenv==0.14.0
pylint==2.6.0
autopep8==1.5.4
Flask-Session==0.3
requests==2.24.0
phonenumbers==8.12.12